+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 3.0.1
 
 - 修复 `Input` 在受控状态下，设置 value 为 null 时报错问题 [#1246](https://github.com/XiaoMi/hiui/issues/1246)
+- 优化 `<Timepicker/>` 时间范围选择器，支持字符串格式的 value [#1245](https://github.com/XiaoMi/hiui/issues/1245)
+- 修复 `Select` 设置 autoload 为 true 时无限的调用接口问题 [#1249](https://github.com/XiaoMi/hiui/issues/1249)
 
 ## 3.0.0